Added file extension for deliverables.

+++ db/migrate/20110704020002_create_main_course.rb
@@ -9,6 +9,6 @@ class CreateMainCourse < ActiveRecord::Migration
   end
 
   def down
-    Course.main.destroy
+    Course.main && Course.main.destroy
   end
 end
